+++ tests/mail-test.js
@@ -5,9 +5,12 @@ const assert = require('assert');
 
 const sandbox = require('sinon').createSandbox();
 
-const Mail = require('./../lib/mail');
+const { Mail, MailError } = require('../lib');
 
-const MailError = require('./../lib/mail-error');
+const JANIS_MAILING_SERVICE = 'mailing';
+const JANIS_MAILING_NAMESPACE = 'email';
+const JANIS_MAILING_METHOD = 'create';
+let headers = {};
 
 describe('Mail', () => {
 
@@ -23,27 +26,70 @@ describe('Mail', () => {
 		};
 	};
 
+	const session = {
+		authenticationData: {
+			userId: '5d1fc1eeb5b68406e0487a06',
+			userIsDev: true,
+			clientId: '5d1fc1eeb5b68406e0487a07',
+			clientCode: 'fizzmodarg',
+			profileId: '5d1fc1eeb5b68406e0487a08',
+			permissions: []
+		},
+		getSessionInstance: TheClass => {
+			const instance = new TheClass();
+			instance.session = session;
+			return instance;
+		}
+	};
+
 	describe('Send', () => {
 
 		describe('Mail structure required validations', () => {
 
-			it('Should throw when the value templateCode is missing', () => {
+			it('Should throw when the values templateCode and body are missing', () => {
 
 				const mail = new Mail();
 
-				assert.rejects(mail.setTo('string').send(), prepareErrorData(
+				assert.rejects(mail.send(), prepareErrorData(
 					MailError.codes.REQUIRED_FIELD_MISSING,
-					'Empty field: templateCode property is required'
+					'Empty field: templateCode or body property is required'
 				));
 			});
 
-			it('Should throw when the value to is missing', () => {
+			it('Should throw when the value to is missing and body is set', () => {
 
 				const mail = new Mail();
 
-				assert.rejects(mail.setTemplateCode('string').send(), prepareErrorData(
+				assert.rejects(mail.setBody('string').send(), prepareErrorData(
 					MailError.codes.REQUIRED_FIELD_MISSING,
-					'Empty field: to property is required'
+					'Empty field: to property is required when templateCode is not set'
+				));
+			});
+
+			it('Should throw when the value clientCode is missing and the session is not set', () => {
+
+				const mail = new Mail();
+
+				assert.rejects(mail.setBody('string')
+					.setTo('example@example.com')
+					.setSubject('crazy subject')
+					.send(),
+				prepareErrorData(
+					MailError.codes.REQUIRED_FIELD_MISSING,
+					'Empty field: clientCode or SessionInstance property is required to send a email'
+				));
+			});
+
+			it('Should throw when the value subject is missing and body is set', () => {
+
+				const mail = new Mail();
+
+				assert.rejects(mail.setBody('string')
+					.setTo('example@example.com')
+					.send(),
+				prepareErrorData(
+					MailError.codes.REQUIRED_FIELD_MISSING,
+					'Empty field: subject property is required when templateCode is not set'
 				));
 			});
 		});
@@ -55,7 +101,9 @@ describe('Mail', () => {
 				const mail = new Mail();
 
 				const mailData = mail.setTo({ wrongFiel: 'string' })
-					.setTemplateCode('template-code')
+					.setBody('body string')
+					.setClientCode('clientCode')
+					.setSubject('crazy subject')
 					.send();
 
 				assert.rejects(mailData, prepareErrorData(MailError.codes.INVALID_FIELD_TYPE, 'Invalid mail: to property must be an array or string'));
@@ -65,8 +113,8 @@ describe('Mail', () => {
 
 				const mail = new Mail();
 
-				const mailData = mail.setTo('string')
-					.setTemplateCode({ templateCode: 'template-code' })
+				const mailData = mail.setTemplateCode({ templateCode: 'template-code' })
+					.setClientCode('clientCode')
 					.send();
 
 				assert.rejects(mailData, prepareErrorData(MailError.codes.INVALID_FIELD_TYPE, 'Invalid mail: templateCode property must be a string or number'));
@@ -78,6 +126,7 @@ describe('Mail', () => {
 
 				const mailData = mail.setTo('string')
 					.setTemplateCode('template-code')
+					.setClientCode('client-code')
 					.setCC({ CC: 'cc addresses' })
 					.send();
 
@@ -90,6 +139,7 @@ describe('Mail', () => {
 
 				const mailData = mail.setTo('string')
 					.setTemplateCode('template-code')
+					.setClientCode('client-code')
 					.setBCC({ BCC: 'bcc addresses' })
 					.send();
 
@@ -102,6 +152,7 @@ describe('Mail', () => {
 
 				const mailData = mail.setTo('string')
 					.setTemplateCode('template-code')
+					.setClientCode('client-code')
 					.setReplyTo({ replyTo: 'replyTo addresses' })
 					.send();
 
@@ -114,6 +165,7 @@ describe('Mail', () => {
 
 				const mailData = mail.setTo('string')
 					.setTemplateCode('template-code')
+					.setClientCode('client-code')
 					.setEntity({ entity: 'entity addresses' })
 					.send();
 
@@ -126,6 +178,7 @@ describe('Mail', () => {
 
 				const mailData = mail.setTo('string')
 					.setTemplateCode('template-code')
+					.setClientCode('client-code')
 					.setEntityId({ entityId: 'entityId addresses' })
 					.send();
 
@@ -138,6 +191,7 @@ describe('Mail', () => {
 
 				const mailData = mail.setTo('string')
 					.setTemplateCode('template-code')
+					.setClientCode('client-code')
 					.setSubject({ subject: 'subject addresses' })
 					.send();
 
@@ -150,10 +204,37 @@ describe('Mail', () => {
 
 				const mailData = mail.setTo('string')
 					.setTemplateCode('template-code')
+					.setClientCode('client-code')
 					.setData('data in string')
 					.send();
 
 				assert.rejects(mailData, prepareErrorData(MailError.codes.INVALID_FIELD_TYPE, 'Invalid mail: data property should be a object'));
+			});
+
+			it('Should throw when the value of \'clientCode\' is not a string', () => {
+
+				const mail = new Mail();
+
+				const mailData = mail.setTo('string')
+					.setTemplateCode('template-code')
+					.setEntity({ entity: 'entity addresses' })
+					.setClientCode({ entity: 'entity addresses' })
+					.send();
+
+				assert.rejects(mailData, prepareErrorData(MailError.codes.INVALID_FIELD_TYPE, 'Invalid mail: clientCode property must be a string'));
+			});
+
+			it('Should throw when the value of \'body\' is not a string', () => {
+
+				const mail = new Mail();
+
+				const mailData = mail.setTo('string')
+					.setBody({ entity: 'entity addresses' })
+					.setClientCode('clientCode')
+					.setSubject('crazy subject')
+					.send();
+
+				assert.rejects(mailData, prepareErrorData(MailError.codes.INVALID_FIELD_TYPE, 'Invalid mail: body property must be a string'));
 			});
 		});
 
@@ -169,24 +250,89 @@ describe('Mail', () => {
 				sandbox.stub(MicroserviceCall.prototype, 'post').resolves(microserviceCallError);
 
 				const mail = new Mail();
-				const mailData = mail.setTo('string')
-					.setTemplateCode('template-code')
-					.send();
+				mail.session = session;
 
-				assert.rejects(mailData, microserviceCallError);
+				await assert.rejects(() => mail
+					.setTo('example@example.com')
+					.setTemplateCode('template-code')
+					.send(), microserviceCallError);
+
+				sandbox.assert.calledOnce(MicroserviceCall.prototype.post);
+				sandbox.assert.calledWithExactly(MicroserviceCall.prototype.post,
+					JANIS_MAILING_SERVICE,
+					JANIS_MAILING_NAMESPACE,
+					JANIS_MAILING_METHOD,
+					{ templateCode: 'template-code', to: 'example@example.com' },
+					headers
+				);
 			});
 
-			it('Should send the email', async () => {
+			it('Should send the email with a templateCode', async () => {
 
 				sandbox.stub(MicroserviceCall.prototype, 'post').resolves({ body: { id: '5de565c07de99000110dcdef' } });
 
 				const mail = new Mail();
-				const mailData = mail.setTo('string')
-					.setTemplateCode('template-code')
+				mail.session = session;
+				const mailData = mail.setTemplateCode('template-code')
 					.send();
 
 
 				assert.deepStrictEqual(await mailData, { id: '5de565c07de99000110dcdef' });
+
+				sandbox.assert.calledOnce(MicroserviceCall.prototype.post);
+				sandbox.assert.calledWithExactly(MicroserviceCall.prototype.post,
+					JANIS_MAILING_SERVICE,
+					JANIS_MAILING_NAMESPACE,
+					JANIS_MAILING_METHOD,
+					{ templateCode: 'template-code' },
+					headers
+				);
+			});
+
+			it('Should send the email with a body', async () => {
+
+				sandbox.stub(MicroserviceCall.prototype, 'post').resolves({ body: { id: '5de565c07de99000110dcdef' } });
+
+				const mail = new Mail();
+				mail.session = session;
+				const mailData = mail.setBody('body string to send')
+					.setTo('example@example.com')
+					.setSubject('crazy subject')
+					.send();
+
+				assert.deepStrictEqual(await mailData, { id: '5de565c07de99000110dcdef' });
+
+				sandbox.assert.calledOnce(MicroserviceCall.prototype.post);
+				sandbox.assert.calledWithExactly(MicroserviceCall.prototype.post,
+					JANIS_MAILING_SERVICE,
+					JANIS_MAILING_NAMESPACE,
+					JANIS_MAILING_METHOD,
+					{ body: 'body string to send', to: 'example@example.com', subject: 'crazy subject' },
+					headers
+				);
+			});
+
+			it('Should send the email with a clientCode', async () => {
+
+				sandbox.stub(MicroserviceCall.prototype, 'post').resolves({ body: { id: '5de565c07de99000110dcdef' } });
+
+				const mail = new Mail();
+				const mailData = mail.setTemplateCode('template-code')
+					.setClientCode('clientCode')
+					.send();
+
+				assert.deepStrictEqual(await mailData, { id: '5de565c07de99000110dcdef' });
+
+				headers = { 'janis-client': 'clientCode' };
+
+				sandbox.assert.calledOnce(MicroserviceCall.prototype.post);
+				sandbox.assert.calledWithExactly(MicroserviceCall.prototype.post,
+					JANIS_MAILING_SERVICE,
+					JANIS_MAILING_NAMESPACE,
+					JANIS_MAILING_METHOD,
+					{ templateCode: 'template-code' },
+					headers
+				);
 			});
 		});
 	});